+++ include/libpldm/utils.h
@@ -24,15 +24,9 @@ struct variable_field {
 	size_t length;
 };
 
-/** @brief Convert ver32_t to string
- *  @param[in] version - Pointer to ver32_t
- *  @param[out] buffer - Pointer to the buffer
- *  @param[in] buffer_size - Size of the buffer, up to SSIZE_MAX
- *  @return The number of characters written to the buffer (excluding the null
- * byte). The converted string may be truncated, and truncation is not
- * considered an error. The result is negative if invalid arguments are supplied
- * (NULL values for required pointers or the buffer size is beyond a
- *  representable range).
+/*
+ * The canonical definition is in base.h. These declarations remain  for
+ * historical compatibility
  */
 ssize_t ver2str(const ver32_t *version, char *buffer, size_t buffer_size);
 ssize_t pldm_base_ver2str(const ver32_t *version, char *buffer,